added edit feature

+++ server.js
@@ -1,6 +1,8 @@
 const admin = require('firebase-admin');
 const serviceAccount = require('./path/to/serviceAccountKey.json');
 
+const collection='url';
+
 if (!admin.apps.length) {
   admin.initializeApp({
       credential: admin.credential.cert(serviceAccount)
@@ -18,20 +20,29 @@ function encryptingURL(longUrl)
   let hash = crypto.createHash('md5').update(name).digest('hex');
   return hash.slice(0,4);
 }
-
+function encodingURL(longUrl)
+{
+  return Buffer.from(longUrl).toString('base64')
+}
+function decodingURL(longUrl)
+{
+  let buff = Buffer.from(longUrl, 'base64');
+  let text = buff.toString('ascii');
+  return text
+}
 async function addDocument(data,id) 
 {
   // converting longurl to base64
   let longUrl = data.original_url;
-  let encodedLongUrl =Buffer.from(longUrl).toString('base64');
+  let encodedLongUrl =encodingURL(longUrl);
   data.original_url=encodedLongUrl;
 // console.log('"' + data + '" converted to Base64 is "' + base64data + '"');
-  const res = await db.collection('url').doc(id).set(data);
+  const res = await db.collection(collection).doc(id).set(data);
 }
 
 async function getOne(req) 
 {
-  const col = db.collection('url').doc(req);
+  const col = db.collection(collection).doc(req);
   const doc= await col.get();
   await col.update({clicks: Number(doc.data().clicks)+1});
   let result;
@@ -39,13 +50,11 @@ async function getOne(req)
   {
     //decode the url
     let data = doc.data().original_url;
-    let buff = Buffer.from(data, 'base64');
-    let text = buff.toString('ascii');
-    result=text;
+    result= decodingURL(data);
   }
   else
   {
-    const res = await col.delete();
+    await col.delete();
   }
   return result;
 }
@@ -54,26 +63,33 @@ async function getOneByPass(req)
 {
   const id=req.body.custom_url;
   const pass=req.body.pass;
-  const col = db.collection('url').doc(id);
+  const col = db.collection(collection).doc(id);
   const doc= await col.get();
   let result=new Object();
   if(doc.data().expiration_date>new Date().getTime()&&doc.data().pass_custom_url==pass)
   {
     //decode the url
     let data = doc.data().original_url;
-    let buff = Buffer.from(data, 'base64');
-    let text = buff.toString('ascii');
-    result.longUrl=text;
+    result.longUrl=decodingURL(data);
     result.shortUrl=doc.data().custom_url;
     result.expiry=doc.data().expiration_date;
     result.clicks=doc.data().clicks;
     return result;
   }
-  else
+  else if(doc.data().expiration_date<new Date().getTime())
   {
-    const res = await col.delete();
+    await col.delete();
     throw new EvalError;
   }
 }
 
-module.exports={addDocument,getOne,encryptingURL,getOneByPass};
+async function updateData(req)
+{
+  const id=req.body.custom_url;
+  const longUrl=encodingURL(req.body.original_url);
+  const pass=req.body.pass;
+  const col = db.collection(collection).doc(id);
+  await col.update({original_url: longUrl,pass_custom_url:pass});
+}
+
+module.exports={addDocument,getOne,encryptingURL,getOneByPass,updateData};
